requirements.txt
flask
pony
# deal with problem: nosetests import ERROR 
nose-pathmunge